mastermind-fixes/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-dotenv
anthropic
google-generativeai
openai
httpx
pydantic
python-jose[cryptography]
groq
pywebpush
resend
